CODE_OF_CONDUCT.md
Task 4: Add a code of conduct
A code of conduct helps set the ground rules for the behavior of your project's participants. It defines standards for how to engage in a community.

GitHub provides templates for common codes of conduct to help you quickly add one to your project. To add a code of conduct to your project, complete the following steps:

Add a new file named CODE_OF_CONDUCT.md to the root folder of the repository with the "Contributor Covenant" template.

Go to your repositories homepage and check if the file has been created.