CONTRIBUTING.md
Task 5: Add contribution guidelines
The contribution guidelines tell project participants how to contribute to the project. To add contributions guidelines, complete the following steps:

Create a new file named CONTRIBUTING.md in the root directory of the repository with the following information:

1
All contributions, bug reports, bug fixes, documentation improvements, enhancements, and ideas are welcome.
Copied!
Optionally, you can review the following guides for examples of contribution guidelines and update this file.

Contributing to Legit Info, a Call for Code for Racial Justice Project
Contributing to OpenEEW
Contributing to Atom
How to contribute to Ruby on Rails
Commit the file into the main branch and check if it is listed on the homepage of the repository.